% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Markov.dir.info-package.R
\docType{package}
\name{Markov.dir.info-package}
\alias{Markov.dir.info}
\alias{Markov.dir.info-package}
\title{Markov.dir.info: Directed Information estimation for Markov chains}
\description{
The package allows user to estimate directed information between nodes of multivariate Markov process.
}
\author{
\strong{Maintainer}: Jacek Zgud \email{jacek.zgud2@gmail.com}

}
\keyword{internal}
