% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovFiltering.R
\name{markov.sim.filt}
\alias{markov.sim.filt}
\title{Simulate obj with marginalized variables.}
\usage{
markov.sim.filt(obj, sim.length, target = c("Y"), condition_set = NULL)
}
\arguments{
\item{obj}{A MarkovProcess object.}

\item{sim.length}{Length of the simulation.}

\item{target}{A vector of variable names to be simulated.}

\item{condition_set}{A vector of variable names to condition by (not implemented).}
}
\value{
Markov filtering simulation.
}
\description{
Simulate obj with marginalized variables.
}
