% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateChain.R
\name{markov.sim}
\alias{markov.sim}
\title{Calculate simulation of markov process, given relevant MarkovProcess class object.
@param obj A MarkovClass object.
@param sim.length A Length of process to be simulated.
@returns  Simulation of markov process of length m.
@export}
\usage{
markov.sim(obj, sim.length)
}
\description{
Calculate simulation of markov process, given relevant MarkovProcess class object.
@param obj A MarkovClass object.
@param sim.length A Length of process to be simulated.
@returns  Simulation of markov process of length m.
@export
}
