% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StationaryDistribution.R
\name{prob_transition}
\alias{prob_transition}
\title{Calculate probability of given configuration at time (t) given configuration at (t-1).}
\usage{
prob_transition(x, names, ParentStructure, TransitionProbabilities)
}
\arguments{
\item{x}{A vector consisting of stacked single configurations at time (t) and (t-1).}

\item{names}{A vector containing node names.}

\item{ParentStructure}{A (d x d) matrix defining parental relations where rows represent parents.}

\item{TransitionProbabilities}{A 3 dimensional array of conditional probabilities.}
}
\value{
Probability of transition.
}
\description{
Calculate probability of given configuration at time (t) given configuration at (t-1).
}
