% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginalRunner.R
\name{simulate.marginalized}
\alias{simulate.marginalized}
\title{Simulate Markov filtering procedure.}
\usage{
simulate.marginalized(obj, target = c(obj@node_names[1]), sim.length = 1000)
}
\arguments{
\item{obj}{A MarkovProcess object.}

\item{target}{A vector of node names to simulate}

\item{sim.length}{Length of simulation to perform.}
}
\value{
A MarkovProcess object.
}
\description{
Simulate Markov filtering procedure.
}
