% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationRunner.R
\name{simulate}
\alias{simulate}
\title{Simulate Markov process}
\usage{
simulate(obj, sim.length = 1000)
}
\arguments{
\item{obj}{A MarkovProcess class.}

\item{sim.length}{Length of simulation.}
}
\description{
Simulate Markov process
}
