% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StationaryDistribution.R
\name{stationary.probability}
\alias{stationary.probability}
\title{Calculate stationary distribution for MarkovProcess object.}
\usage{
stationary.probability(obj)
}
\arguments{
\item{obj}{A MarkovProcess object.}
}
\value{
A MarkovProcess object with populated slots assosiated with stationary distribution.
}
\description{
Calculate stationary distribution for MarkovProcess object.
}
