% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StationaryDistribution.R
\name{trans.matrix}
\alias{trans.matrix}
\title{Calculates transition matrix for MarkovProcess class}
\usage{
trans.matrix(obj)
}
\arguments{
\item{obj}{A MarkovProcess object.}
}
\value{
A MarkovProcess object with some slots populated.
}
\description{
Calculates transition matrix for MarkovProcess class
}
